+++ src/app/features/designer/designer-demo.routes.ts
@@ -1,0 +1,9 @@
+import { Routes } from '@angular/router';
+import { DesignerDemoComponent } from './designer-demo.component';
+
+export const routes: Routes = [
+  {
+    path: '',
+    component: DesignerDemoComponent
+  }
+];